+++ Cargo.toml
@@ -15,6 +15,7 @@ lto = true      # link with link time optimization (lto).
 debug = true    # enable debugging in release mode.
 
 [dependencies]
+cty = "0.2.2"
 cortex-m = "0.7.7"
 cortex-m-rt = "0.7.3"
 cortex-m-semihosting = "0.5.0"
@@ -28,4 +29,4 @@ features = ["mcu-STM32L010RBTx","stm32l0x1","rt", "disable-linker-script"]
 # mcu-STM32L010RBTx = ["lqfp64", "io-STM32L071", "eeprom-512", "flash-128", "ram-20"]
 
 [build.dependencies]
-bindgen = "0.66.1"
+cc = "1.0.79"
