+++ build.rs
@@ -1,0 +1,7 @@
+extern crate cc;
+
+fn main() {
+    cc::Build::new()
+        .file("src/doubler.c")
+        .compile("libdoubler.a");
+}
