+++ src/doubler.c
@@ -1,0 +1,6 @@
+#include "doubler.h"
+extern const int FACTOR;
+
+int doubler(int x) {
+        return x * FACTOR;
+}
