+++ src/doubler.h
@@ -1,0 +1,3 @@
+const int FACTOR = 2;
+
+int doubler(int x);
