+++ src/main.rs
@@ -6,30 +6,16 @@ extern crate panic_halt;
 extern crate stm32l0;
 
 use cortex_m_rt::entry;
-use stm32l0xx_hal::{pac, prelude::*, rcc::Config};
+
+#[link(name = "doubler")]
+extern "C" {
+    fn doubler(x: i32) -> i32;
+}
 
 #[entry]
-fn main() -> ! {
-    let cp = cortex_m::Peripherals::take().unwrap();
-    let dp = pac::Peripherals::take().unwrap();
-
-    // Configure the clock.
-    let mut rcc = dp.RCC.freeze(Config::hsi16());
-
-    // Acquire the GPIOA peripheral. This also enables the clock for GPIOA in the RCC register.
-    let gpioa = dp.GPIOA.split(&mut rcc);
-
-    // Configure LED2 as output.
-    let mut led2 = gpioa.pa5.into_push_pull_output();
-
-    // Get the delay provider.
-    let mut delay = cp.SYST.delay(rcc.clocks);
-
+unsafe fn main() -> ! {
+    let result = unsafe { doubler(5) };
     loop {
-        led2.set_high().unwrap();
-        delay.delay_ms(300_u16);
-
-        led2.set_low().unwrap();
-        delay.delay_ms(300_u16);
+        cortex_m_semihosting::hprintln!("Doubler result: {}", result);
     }
 }
